% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_bar.R
\name{distribution_bar}
\alias{distribution_bar}
\title{Distribution Plot}
\usage{
distribution_bar(data, x, title = NULL, x_label = NULL, y_label = NULL)
}
\arguments{
\item{data}{data_frame A data frame or a tibble.}

\item{x}{unquoted column name to plot on the x-axis from data data_frame or tibble}
}
\value{
A ggplot object
}
\description{
A function for generating a bar for the distribution of a variable based
on its frequency,
}
