% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{quality}
\alias{quality}
\title{Transform quality values}
\usage{
quality(data)
}
\arguments{
\item{data}{A dataframe containing a 'quality' column.}
}
\value{
The input dataframe with the 'quality' column transformed.
}
\description{
This function transforms the 'quality' column in a dataframe to categorical values "bad" or "good".
}
\examples{
data <- data.frame(quality = c(3, 4, 5, 6, 7))
quality(data)

}
