% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Function to plot a scatter plot}
\usage{
scatter_plot(data, x_var, y_var, title = NULL, x_label = NULL, y_label = NULL)
}
\arguments{
\item{data}{A data frame containing the prediction results}

\item{x_var}{the x-axis chosen to make the plot}

\item{y_var}{the y-axis chosen to make the plot}

\item{x_label}{The name for the x-axis label}

\item{y_label}{The name for the y-axis label}
}
\value{
A ggplot object
}
\description{
Function to plot a scatter plot
}
